feat: move feed to main page
