--- app/profile/loading.tsx
@@ -1,3 +1,0 @@
-const Loading = () => <h1>loading...</h1>;
-
-export default Loading;
