+++ app/profile/page.tsx
@@ -1,9 +1,11 @@
 import { collection, query, where, getDocs } from "firebase/firestore";
 import { unstable_getServerSession } from "next-auth";
 
+import FeedPage from "../../components/feed/FeedPage";
 import { authOptions } from "../../pages/api/auth/[...nextauth]";
-import FeedPage from "../feed/FeedPage";
 import db from "../firebase";
+
+import type { FeedItemContent } from "../../components/feed/FeedPage";
 
 // TODO SOPHIE FIX THIS (make it a utility function so we can use it in other places)
 const getUser = async () => {
@@ -17,11 +19,6 @@ const getUser = async () => {
   return user;
 };
 
-export type FeedItemRename = {
-  id: string;
-  data: { username: string; content: string };
-};
-
 async function getData() {
   const user = await getUser();
   // change to profile's username
@@ -32,7 +29,7 @@ async function getData() {
   const qSnapshot = await getDocs(q);
 
   return qSnapshot.docs.map(
-    (doc) => ({ id: doc.id, data: doc.data() } as FeedItemRename)
+    (doc) => ({ id: doc.id, data: doc.data() } as FeedItemContent)
   );
 }
